+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Infrastructure/Migrations/20250513070441_AddTagGroupColour.cs
@@ -1,0 +1,44 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace FF.Articles.Backend.Contents.API.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddTagGroupColour : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "TagColour",
+                schema: "Contents",
+                table: "Tag",
+                type: "character varying(128)",
+                maxLength: 128,
+                nullable: true);
+
+            migrationBuilder.AddColumn<string>(
+                name: "TagGroup",
+                schema: "Contents",
+                table: "Tag",
+                type: "character varying(128)",
+                maxLength: 128,
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "TagColour",
+                schema: "Contents",
+                table: "Tag");
+
+            migrationBuilder.DropColumn(
+                name: "TagGroup",
+                schema: "Contents",
+                table: "Tag");
+        }
+    }
+}