+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Dtos/TagDto.cs
@@ -3,5 +3,7 @@ public class TagDto
 {
     public long TagId { get; set; }
     public string TagName { get; set; }
+    public string? TagGroup { get; set; }
+    public string? TagColour { get; set; }
 
 }